# Try to optimise bracket matching.