Related to ipython/ipython#13731

Finding matching brackets in in some cases, this is one of the slowest operation.

Test script:

> import pandas.core.series as ps
> from asttokens import ASTTokens
> from pathlib import Path
>
> text = Path(ps.__file__).read_text()
>
> for i in range(10):
>     ASTTokens(text, True)

$ time python opt.py
python opt.py  1.91s user 0.14s system 122% cpu 1.674 total

vs master branch

python opt.py  2.20s user 0.13s system 119% cpu 1.957 total

In the tree-visit step for the specific above example this reduces the
time matching brackets from 2/3 of the function time (it is call in 2
places), to arround 20%, so about a x3 speed increase.

I want to make sure you are ok with this approach before going further
as I have spent already quite a bit of time on this

The line profiling also gives the following on master branch:

    Total time: 2.53663 s
    Function: _visit_after_children at line 81

    Line #      Hits         Time  Per Hit   % Time  Line Contents
    ==============================================================
        81                                             @profile
        82                                             def _visit_after_children(self, node, parent_token, token):
        83                                               # type: (AstNode, Optional[util.Token], Optional[util.Token]) -> None
        84                                               # This processes the node generically first, after all children have been processed.
        85
        86                                               # Get the first and last tokens that belong to children. Note how this doesn't assume that we
        87                                               # iterate through children in order that corresponds to occurrence in source code. This
        88                                               # assumption can fail (e.g. with return annotations).
        89     69940      18300.0      0.3      0.7      first = token
        90     69940      18611.0      0.3      0.7      last = None
        91    139870     365826.0      2.6     14.4      for child in cast(Callable, self._iter_children)(node):
        92     69930      26697.0      0.4      1.1        if not first or child.first_token.index < first.index:
        93      2240        658.0      0.3      0.0          first = child.first_token
        94     69930      21753.0      0.3      0.9        if not last or child.last_token.index > last.index:
        95     66260      19586.0      0.3      0.8          last = child.last_token
        96
        97                                               # If we don't have a first token from _visit_before_children, and there were no children, then
        98                                               # use the parent's token as the first token.
        99     69940      19504.0      0.3      0.8      first = first or parent_token
       100
       101                                               # If no children, set last token to the first one.
       102     69940      18191.0      0.3      0.7      last = last or first
       103
       104                                               # Statements continue to before NEWLINE. This helps cover a few different cases at once.
       105     69940      39496.0      0.6      1.6      if util.is_stmt(node):
       106     10310      49476.0      4.8      2.0        last = self._find_last_in_stmt(cast(util.Token, last))
       107
       108                                               # Capture any unmatched brackets.
       109     69940    1532993.0     21.9     60.4      first, last = self._expand_to_matching_pairs(cast(util.Token, first), cast(util.Token, last), node)
       110
       111                                               # Give a chance to node-specific methods to adjust.
       112     69940     183633.0      2.6      7.2      nfirst, nlast = self._methods.get(self, node.__class__)(node, first, last)
       113
       114     69940      24949.0      0.4      1.0      if (nfirst, nlast) != (first, last):
       115                                                 # If anything changed, expand again to capture any unmatched brackets.
       116      7580     154927.0     20.4      6.1        nfirst, nlast = self._expand_to_matching_pairs(nfirst, nlast, node)
       117
       118     69940      21861.0      0.3      0.9      node.first_token = nfirst
       119     69940      20169.0      0.3      0.8      node.last_token = nlast


And on this PR:

    Total time: 1.08657 s
    Function: _visit_after_children at line 81

    Line #      Hits         Time  Per Hit   % Time  Line Contents
    ==============================================================
        81                                             @profile
        82                                             def _visit_after_children(self, node, parent_token, token):
        83                                               # type: (AstNode, Optional[util.Token], Optional[util.Token]) -> None
        84                                               # This processes the node generically first, after all children have been processed.
        85
        86                                               # Get the first and last tokens that belong to children. Note how this doesn't assume that we
        87                                               # iterate through children in order that corresponds to occurrence in source code. This
        88                                               # assumption can fail (e.g. with return annotations).
        89     69940      18259.0      0.3      1.7      first = token
        90     69940      18593.0      0.3      1.7      last = None
        91    139870     366285.0      2.6     33.7      for child in cast(Callable, self._iter_children)(node):
        92     69930      27234.0      0.4      2.5        if not first or child.first_token.index < first.index:
        93      2240        730.0      0.3      0.1          first = child.first_token
        94     69930      22926.0      0.3      2.1        if not last or child.last_token.index > last.index:
        95     66260      20105.0      0.3      1.9          last = child.last_token
        96
        97                                               # If we don't have a first token from _visit_before_children, and there were no children, then
        98                                               # use the parent's token as the first token.
        99     69940      19850.0      0.3      1.8      first = first or parent_token
       100
       101                                               # If no children, set last token to the first one.
       102     69940      19385.0      0.3      1.8      last = last or first
       103
       104                                               # Statements continue to before NEWLINE. This helps cover a few different cases at once.
       105     69940      39364.0      0.6      3.6      if util.is_stmt(node):
       106     10310      60294.0      5.8      5.5        last = self._find_last_in_stmt(cast(util.Token, last))
       107
       108                                               # Capture any unmatched brackets.
       109     69940     195805.0      2.8     18.0      first, last = self._expand_to_matching_pairs_fast(cast(util.Token, first), cast(util.Token, last))
       110
       111                                               # Give a chance to node-specific methods to adjust.
       112     69940     190991.0      2.7     17.6      nfirst, nlast = self._methods.get(self, node.__class__)(node, first, last)
       113
       114     69940      25562.0      0.4      2.4      if (nfirst, nlast) != (first, last):
       115                                                 # If anything changed, expand again to capture any unmatched brackets.
       116      7580      17687.0      2.3      1.6        nfirst, nlast = self._expand_to_matching_pairs_fast(nfirst, nlast)
       117
       118     69940      22287.0      0.3      2.1      node.first_token = nfirst
       119     69940      21214.0      0.3      2.0      node.last_token = nlast

I'm still not super happy with the speed, but we are closer to the speed
of tokenize.

    Line #      Hits         Time  Per Hit   % Time  Line Contents
    ==============================================================
        53                                             @profile
        54                                             def __init__(self, source_text, parse=False, tree=None, filename='<unknown>'):
        55                                               # type: (Any, bool, Optional[Module], str) -> None
        56                                               # FIXME: Strictly, the type of source_type is one of the six string types, but hard to specify with mypy given
        57                                               # https://mypy.readthedocs.io/en/stable/common_issues.html#variables-vs-type-aliases
        58
        59        10         16.0      1.6      0.0      self._filename = filename
        60        10     139398.0  13939.8      2.8      self._tree = ast.parse(source_text, filename) if parse else tree
        61
        62                                               # Decode source after parsing to let Python 2 handle coding declarations.
        63                                               # (If the encoding was not utf-8 compatible, then even if it parses correctly,
        64                                               # we'll fail with a unicode error here.)
        65        10         55.0      5.5      0.0      source_text = six.ensure_text(source_text)
        66
        67        10          6.0      0.6      0.0      self._text = source_text
        68        10      25998.0   2599.8      0.5      self._line_numbers = LineNumbers(source_text)
        69
        70                                               # Tokenize the code.
        71        10    2646430.0 264643.0     53.4      self._tokens = list(self._generate_tokens(source_text))
        72        10     146805.0  14680.5      3.0      self._tp = make_tree(self._tokens)
        73
        74                                               # Extract the start positions of all tokens, so that we can quickly map positions to tokens.
        75        10      12838.0   1283.8      0.3      self._token_offsets = [tok.startpos for tok in self._tokens]
        76
        77        10         10.0      1.0      0.0      if self._tree:
        78        10    1988787.0 198878.7     40.1        self.mark_tokens(self._tree)